Move MqttServiceInfo to init.py
